# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brand Analysis API",
    "description": "品牌分析API服务，为dashboard提供数据接口",
    "version": "0.1.0"
  },
  "paths": {
    "/api/analysis/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Brand",
        "description": "执行品牌分析.",
        "operationId": "analyze_brand_api_analysis_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/recognize-brand": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Recognize Brand",
        "description": "品牌识别.",
        "operationId": "recognize_brand_api_analysis_recognize_brand_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandRecognitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandRecognitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/results/{result_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis Result",
        "description": "获取分析结果.",
        "operationId": "get_analysis_result_api_analysis_results__result_id__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/history": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis History",
        "description": "获取分析历史记录.",
        "operationId": "get_analysis_history_api_analysis_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Analysis History Api Analysis History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "brand_name": {
            "type": "string",
            "title": "Brand Name",
            "description": "品牌名称"
          },
          "analysis_type": {
            "$ref": "#/components/schemas/AnalysisType",
            "description": "分析类型"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params",
            "description": "分析参数"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider",
            "description": "LLM提供商",
            "default": "openai"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "模型名称",
            "default": "gpt-3.5-turbo"
          }
        },
        "type": "object",
        "required": [
          "brand_name",
          "analysis_type"
        ],
        "title": "AnalysisRequest",
        "description": "分析请求模型."
      },
      "AnalysisResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "是否成功"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "时间戳"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "timestamp"
        ],
        "title": "AnalysisResponse",
        "description": "分析响应模型."
      },
      "AnalysisType": {
        "type": "string",
        "enum": [
          "brand_recognition",
          "sentiment_analysis",
          "competitive_analysis",
          "mention_analysis"
        ],
        "title": "AnalysisType",
        "description": "分析类型枚举."
      },
...
      "BrandRecognitionRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "要分析的文本"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "上下文信息"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider",
            "description": "LLM提供商",
            "default": "openai"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "模型名称",
            "default": "gpt-3.5-turbo"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "BrandRecognitionRequest",
        "description": "品牌识别请求模型."
      },
      "BrandRecognitionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "是否成功"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandRecognitionResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "识别结果"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "时间戳"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "timestamp"
        ],
        "title": "BrandRecognitionResponse",
        "description": "品牌识别响应模型."
      },
      "BrandRecognitionResult": {
        "properties": {
          "brands_found": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Brands Found",
            "description": "发现的品牌列表"
          },
          "mentions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Mentions",
            "description": "品牌提及详情"
          },
          "sentiment_scores": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Scores",
            "description": "情感评分"
          }
        },
        "type": "object",
        "required": [
          "brands_found",
          "mentions"
        ],
        "title": "BrandRecognitionResult",
        "description": "品牌识别结果模型."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}